Unsigned/signed warnings
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS@1992 212acab6-be3b-0410-9dea-997c60f758d6

+++ trunk/Source/script.cpp
@@ -1680,7 +1680,7 @@ int CEXEBuild::doCommand(int which_token, LineParser &line, FILE *fp, const char
           char *incfile=(char*)malloc(size);
 
           for (unsigned int i=0; i<dirs; i++) {
-            if (size < lstrlen(f)+lstrlen(dir))
+            if (size < (unsigned int)lstrlen(f)+lstrlen(dir))
             {
               free(incfile);
               size+=lstrlen(dir);
